refactor out Greeting so 1 file = 1 component
